+++ src/components/CourseList/CourseItem.scss
@@ -45,12 +45,3 @@
 .lessonCount {
   color: $color-gray;
 }
-
-.tooltip {
-  :global(.tooltip-inner) {
-    max-width: 320px;
-  }
-  img {
-    max-height: 180px;
-  }
-}
