+++ src/components/PlaylistPage/LessonItem.scss
@@ -32,12 +32,3 @@
 .progress {
   opacity: 0.5;
 }
-
-.tooltip {
-  :global(.tooltip-inner) {
-    max-width: 320px;
-  }
-  img {
-    max-height: 180px;
-  }
-}
