+++ src/components/TooltipComponent.scss
@@ -1,0 +1,8 @@
+.tooltip {
+  :global(.tooltip-inner) {
+    max-width: 320px;
+  }
+  img {
+    max-height: 180px;
+  }
+}
